# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Personnal Nutrition",
        "description": "",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/users/get-list": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get users list",
                "description": "Get the app users list",
                "operationId": "a296bdb02454084f2347b059c418f520",
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/admin/users/delete/{identifier}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remove user from the system",
                "description": "Delete a user",
                "operationId": "42dc708b7a6a3699308bbae482b59324",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "The identifier of the user to delete (ID or slug)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted successfully"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "Enter your API token here",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}